(1) fixed MCS, needed for the state encoding. (2) added collapseIsoMorphicSuperType method, using this directly in queried grounded content. (3) state encoding now done. Only adds the features not already there. Feature checking working. Dialogues which should lead to identical contexts do. (4) added support for question marks/intonations that hand the turn over, e.g. “by?”, or “you want?” (5) some of the dialogues we discussed are under babble/data/Domain-Dialogues/shopping-mall-interactional-variation
(6) state encoding class now comes up with a SET of goals states after
processing example dialogues. Each goal is a vector of Integers denoting
the indices of the state vector that need to be 1 for it to count as
goal, with the rest as 0
(7) parsing the dialogues in shopping-mall-interactional-variation,
leads only to ONE goal, because they are interactional variations. So
for our experiment for this paper we really only need to parse one of
these. But I don’t now see why we don’t also include some lexical
variation in a second experiment. All is ready for that too.